requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-groq>=0.1.0
duckduckgo-search>=5.3.0
psycopg2-binary>=2.9.9
streamlit>=1.40.0
python-dotenv>=1.0.1
typing-extensions>=4.12.2
fastapi>=0.115.0
uvicorn[standard]>=0.32.0